requirements.txt
numpy
pandas
matplotlib
scikit-learn
nltk
auto-sklearn
tqdm